+++ webapp/channels/src/actions/channel_actions.ts
@@ -11,10 +11,10 @@ import {PreferenceTypes} from 'mattermost-redux/action_types';
 import * as ChannelActions from 'mattermost-redux/actions/channels';
 import {fetchDeletedPostsIds, postDeleted} from 'mattermost-redux/actions/posts';
 import {savePreferences} from 'mattermost-redux/actions/preferences';
-import {getChannelByName, getUnreadChannelIds, getChannel} from 'mattermost-redux/selectors/entities/channels';
-import {getCurrentChannelId, getMyChannelMemberships} from 'mattermost-redux/selectors/entities/common';
+import {getChannelByName, getUnreadChannelIds, getChannel, getRedirectChannelNameForTeam} from 'mattermost-redux/selectors/entities/channels';
+import {getMyChannelMemberships} from 'mattermost-redux/selectors/entities/common';
 import {getPost} from 'mattermost-redux/selectors/entities/posts';
-import {getCurrentTeamUrl, getCurrentTeamId} from 'mattermost-redux/selectors/entities/teams';
+import {getCurrentTeamUrl, getCurrentTeamId, getCurrentRelativeTeamUrl} from 'mattermost-redux/selectors/entities/teams';
 import {getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
 import type {ActionFuncAsync} from 'mattermost-redux/types/actions';
 
@@ -98,12 +98,18 @@ export function loadChannelsForCurrentUser(): ActionFuncAsync {
     };
 }
 
-export function loadDeletedPosts(lastDisconnectAt: number): ActionFuncAsync {
+export function loadDeletedPosts(channelId: string, lastDisconnectAt: number): ActionFuncAsync {
     return async (dispatch, getState) => {
         const state = getState();
-        const channelId = getCurrentChannelId(state);
 
         const results = await dispatch(fetchDeletedPostsIds(channelId, lastDisconnectAt));
+
+        if ('error' in results) {
+            const currentTeamId = getCurrentTeamId(state);
+            const redirectChannel = getRedirectChannelNameForTeam(state, currentTeamId);
+            const teamUrl = getCurrentRelativeTeamUrl(state);
+            getHistory().push(`${teamUrl}/channels/${redirectChannel}`);
+        }
 
         if (results.data && Array.isArray(results.data)) {
             results.data.forEach((postId) => {