+++ webapp/channels/src/actions/global_actions.tsx
@@ -21,7 +21,7 @@ import {Preferences} from 'mattermost-redux/constants';
 import {appsEnabled} from 'mattermost-redux/selectors/entities/apps';
 import {getCurrentChannelStats, getCurrentChannelId, getMyChannelMember, getRedirectChannelNameForTeam, getChannelsNameMapInTeam, getAllDirectChannels, getChannelMessageCount} from 'mattermost-redux/selectors/entities/channels';
 import {getConfig, isPerformanceDebuggingEnabled} from 'mattermost-redux/selectors/entities/general';
-import {getBool, getIsOnboardingFlowEnabled, isCollapsedThreadsEnabled} from 'mattermost-redux/selectors/entities/preferences';
+import {getBool, getTeamsOrderPreference, isCollapsedThreadsEnabled} from 'mattermost-redux/selectors/entities/preferences';
 import {getCurrentTeamId, getTeam, getMyTeamMember, getTeamMemberships, getMyKSuites} from 'mattermost-redux/selectors/entities/teams';
 import {getCurrentUser, getCurrentUserId, isFirstAdmin} from 'mattermost-redux/selectors/entities/users';
 import type {ActionFuncAsync, ThunkActionFunc} from 'mattermost-redux/types/actions';
@@ -423,6 +423,29 @@ export async function redirectUserToDefaultTeam() {
     }
 
     getHistory().push('/select_team');
+}
+
+export async function redirectDesktopUserToDefaultTeam() {
+    const state = getState();
+    const locale = getCurrentLocale(state);
+    const teams = getMyKSuites(state);
+    const activeTeams = teams.filter((team) => team.delete_at === 0);
+
+    const orderPreference = getTeamsOrderPreference(state);
+    const orderedTeams = filterAndSortTeamsByDisplayName(activeTeams, locale, orderPreference.value);
+    const newCurrentTeam = orderedTeams[0];
+
+    if (newCurrentTeam) {
+        window.postMessage(
+            {
+                type: 'switch-server',
+                data: newCurrentTeam.display_name,
+            },
+            window.origin,
+        );
+    } else {
+        getHistory().push('/error?type=page_not_found');
+    }
 }
 
 export function joinChannel(channelId: string) {