+++ webapp/channels/src/actions/user_actions.ts
@@ -29,7 +29,7 @@ import {calculateUnreadCount} from 'mattermost-redux/utils/channel_utils';
 
 import {loadCustomEmojisForCustomStatusesByUserIds} from 'actions/emoji_actions';
 import {loadStatusesForProfilesList, loadStatusesForProfilesMap} from 'actions/status_actions';
-import {getDisplayedChannels} from 'selectors/views/channel_sidebar';
+import {getDisplayedChannels, getSidebarGroupChannels} from 'selectors/views/channel_sidebar';
 import store from 'stores/redux_store';
 
 import {Constants, Preferences, UserStatuses} from 'utils/constants';
@@ -295,13 +295,7 @@ export async function loadProfilesForSidebar() {
 }
 
 export const getGMsForLoading = (state: GlobalState) => {
-    // Get all channels visible on the current team which doesn't include hidden GMs/DMs
-    let channels = getDisplayedChannels(state);
-
-    // Make sure we only have GMs
-    channels = channels.filter((channel) => channel.type === General.GM_CHANNEL);
-
-    return channels;
+    return getSidebarGroupChannels(state);
 };
 
 export async function loadProfilesForGM() {