+++ webapp/channels/src/actions/views/drafts.ts
@@ -115,15 +115,17 @@ export function removeDraft(key: string, channelId: string, rootId = ''): Action
             // const connectionId = getConnectionId(getState());
 
             try {
-                console.log(`@debug_draft removing draft from api ~ key: ${key}, channelId: ${channelId}, rootId: ${rootId}`);
-
                 if (draft.timestamp) {
+                    console.log(`[actions/drafts] removeDraft.deleteScheduledDraft key: ${key}, channelId: ${channelId}, rootId: ${rootId}`);
                     await Client4.deleteScheduledDraft(draft.id);
                 } else {
+                    console.log(`[actions/drafts] removeDraft.deleteDraft key: ${key}, channelId: ${channelId}, rootId: ${rootId}`);
                     await Client4.deleteDraft(channelId, rootId);
                 }
             } catch (error) {
-                console.log('@debug_draft Could not delete draft from api', key);
+                console.log(`[actions/drafts] removeDraft.deleteDraft:error key: ${key}`,
+                    `error: ${JSON.stringify(error)}`,
+                );
                 return {
                     data: false,
                     error,
@@ -160,10 +162,22 @@ export function updateDraft(key: string, value: PostDraft|null, rootId = '', sav
         if (scheduleDelete) {
             const newValue = {...updatedValue};
             Reflect.deleteProperty(newValue, 'timestamp');
-            console.log('@debug_draft updateDraft (scheduled)', `key: ${key}`, `id: ${newValue?.id}`, `updatedAt: ${newValue?.updateAt}`, `createdAt: ${newValue?.createAt}`);
+            console.log('[actions/drafts] updateDraft.setGlobalDraft (scheduled)',
+                `key: ${key}`,
+                `id: ${newValue?.id}`,
+                `updatedAt: ${newValue?.updateAt}`,
+                `createdAt: ${newValue?.createAt}`,
+            );
+
             dispatch(setGlobalDraft(key, newValue as PostDraft, false));
         } else {
-            console.log('@debug_draft updateDraft', `key: ${key}`, `id: ${updatedValue?.id}`, `updatedAt: ${updatedValue?.updateAt}`, `createdAt: ${updatedValue?.createAt}`);
+            console.log('[actions/drafts] updateDraft.setGlobalDraft',
+                `key: ${key}`,
+                `id: ${updatedValue?.id}`,
+                `updatedAt: ${updatedValue?.updateAt}`,
+                `createdAt: ${updatedValue?.createAt}`,
+            );
+
             dispatch(setGlobalDraft(key, updatedValue, false));
         }
 
@@ -171,15 +185,21 @@ export function updateDraft(key: string, value: PostDraft|null, rootId = '', sav
             const userId = getCurrentUserId(state);
 
             try {
+                // TODO: remove
                 if (value?.message === '' || value?.message.replace(/\s/g, '').length || (value && value?.fileInfos.length > 0)) {
                     if (value?.message.replace(/\s/g, '').length) {
                         const {id} = await upsertDraft(updatedValue, userId, rootId, scheduleDelete);
-                        updatedValue.id = id;
+                        dispatch(setGlobalDraft(key, {
+                            ...updatedValue,
+                            id,
+                        }, false));
                     } else {
                         //This case is when there is a file attached with no message
                         await upsertDraft({...updatedValue, message: ''}, userId, rootId, scheduleDelete);
                     }
                 }
+
+                // await upsertDraft(updatedValue, userId, rootId, scheduleDelete);
             } catch (error) {
                 return {data: false, error};
             }
@@ -201,15 +221,31 @@ export function upsertScheduleDraft(key: string, value: PostDraft, rootId = ''):
         dispatch(setGlobalItem(key, {message: '', fileInfos: [], uploadsInProgress: []}));
 
         try {
-            console.log('@debug_draft upsertScheduleDraft', `id: ${value?.id}`, `updatedAt: ${value?.updateAt}`, `createdAt: ${value?.createAt}`);
+            console.log('[actions/drafts] upsertScheduleDraft.upsertDraft',
+                `id: ${value?.id}`,
+                `updatedAt: ${value?.updateAt}`,
+                `createdAt: ${value?.createAt}`,
+            );
+
             const {id} = await upsertDraft(value, userId, rootId);
             dispatch(setGlobalItem(`${key}_${id}`, {
                 ...value,
                 id,
             }));
         } catch (error) {
-            console.log('@debug_draft upsertScheduleDraft Error', `id: ${value?.id}`, `updatedAt: ${value?.updateAt}`, `createdAt: ${value?.createAt}`);
+            console.log('[actions/drafts] upsertScheduleDraft.upsertDraft:error',
+                `id: ${value?.id}`,
+                `updatedAt: ${value?.updateAt}`,
+                `createdAt: ${value?.createAt}`,
+                `error: ${JSON.stringify(error)}`,
+            );
             if (activeDraft) {
+                console.log('[actions/drafts] upsertScheduleDraft.upsertDraft:error setGlobalItem',
+                    `id: ${value?.id}`,
+                    `updatedAt: ${value?.updateAt}`,
+                    `createdAt: ${value?.createAt}`,
+                );
+
                 dispatch(setGlobalItem(key, activeDraft));
             }
             return {error};
@@ -240,11 +276,20 @@ function upsertDraft(draft: PostDraft, userId: UserProfile['id'], rootId = '', s
             Reflect.deleteProperty(newDraft, 'timestamp');
         }
 
-        console.log('@debug_draft upsertDraft (scheduleDrat)', `id: ${newDraft?.id}`, `updatedAt: ${draft?.updateAt}`, `createdAt: ${draft?.createAt}`);
+        console.log('[actions/drafts] upsertDraft.updateScheduledDraft',
+            `id: ${newDraft?.id}`,
+            `updatedAt: ${draft?.updateAt}`,
+            `createdAt: ${draft?.createAt}`,
+        );
+
         return Client4.updateScheduledDraft(newDraft);
     }
 
-    console.log('@debug_draft upsertDraft', `id: ${newDraft?.id}`, `updatedAt: ${draft?.updateAt}`, `createdAt: ${draft?.createAt}`);
+    console.log('[actions/drafts] upsertDraft.upsertDraft',
+        `id: ${newDraft?.id}`,
+        `updatedAt: ${draft?.updateAt}`,
+        `createdAt: ${draft?.createAt}`,
+    );
 
     return Client4.upsertDraft(newDraft);
 }
@@ -266,7 +311,17 @@ export function setDraftsTourTipPreference(initializationState: Record<string, b
 
 export function setGlobalDraft(key: string, value: PostDraft|null, isRemote: boolean): ActionFunc {
     return (dispatch) => {
-        dispatch(setGlobalItem(key, value));
+        if (value && !value.message) {
+            value.message = ''; // prevent message to be null and raise an exception
+        }
+        if (value?.message === '' || (value?.message?.replace(/\s/g, '').length) || (value && value?.fileInfos.length > 0)) {
+            if (value?.message.replace(/\s/g, '').length) {
+                dispatch(setGlobalItem(key, value));
+            } else {
+                //This case is when there is a file attached with no message
+                dispatch(setGlobalItem(key, {...value, message: ''}));
+            }
+        }
         dispatch(setGlobalDraftSource(key, isRemote));
         return {data: true};
     };