+++ webapp/channels/src/components/activity_and_insights/insights/least_active_channels/least_active_channels_item/least_active_channels_item.tsx
@@ -94,6 +94,7 @@ const LeastActiveChannelsItem = ({channel, actionCallback}: Props) => {
                 userIds={channel.participants}
                 size='xs'
                 disableProfileOverlay={true}
+                showDeleted={false}
             />
             <ChannelActionsMenu
                 channel={channel}