+++ webapp/channels/src/components/advanced_create_post/advanced_create_post.tsx
@@ -482,9 +482,8 @@ class AdvancedCreatePost extends React.PureComponent<Props, State> {
     };
 
     doSubmit = async (e?: React.FormEvent, isSchedule = false, scheduleUTCTimestamp?: number) => {
-        console.log(`@debug_draft AdvancedCreatePost ~ doSubmit: isSchedule: ${isSchedule}, scheduleUTCTimestamp: ${scheduleUTCTimestamp}`);
+        console.log(`[advanced_create_post ~ draft] doSubmit: isSchedule: ${isSchedule}, scheduleUTCTimestamp: ${scheduleUTCTimestamp}`);
         const channelId = this.props.currentChannel.id;
-        console.log(`@debug_draft AdvancedCreatePost ~ doSubmit: channelId: ${channelId}`);
 
         if (e) {
             e.preventDefault();
@@ -540,6 +539,7 @@ class AdvancedCreatePost extends React.PureComponent<Props, State> {
                 channelId,
                 remote: false,
             };
+            console.log('[advanced_create_post ~ draft] doSubmit.scheduleDraft id: ', updatedDraft.id);
             this.scheduleDraft(updatedDraft);
             return;
         } else if (post.message.indexOf('/') === 0 && !ignoreSlash) {
@@ -603,7 +603,7 @@ class AdvancedCreatePost extends React.PureComponent<Props, State> {
             clearTimeout(this.saveDraftFrame);
         }
 
-        console.log('@debug_draft AdvancedCreatePost ~ doSubmit: removeDraft', channelId);
+        console.log('[advanced_create_post ~ draft] doSubmit: removeDraft', channelId);
         this.isDraftSubmitting = false;
         this.removeDraft(channelId);
     };
@@ -968,7 +968,7 @@ class AdvancedCreatePost extends React.PureComponent<Props, State> {
     };
 
     removeDraft = (channelId = this.props.currentChannel.id) => {
-        console.log('@debug_draft advanced_create_post ~ removeDraft ~', StoragePrefixes.DRAFT + channelId);
+        console.log('[advanced_create_post ~ draft] removeDraft ~', StoragePrefixes.DRAFT + channelId);
         this.props.actions.setDraft(StoragePrefixes.DRAFT + channelId, null, channelId);
         this.draftsForChannel[channelId] = null;
     };