+++ webapp/channels/src/components/main_menu/main_menu.tsx
@@ -101,7 +101,7 @@ export class MainMenu extends React.PureComponent<Props> {
     };
 
     handleEmitUserLoggedOutEvent = (): void => {
-        GlobalActions.emitUserLoggedOutEvent();
+        GlobalActions.emitUserLoggedOutEvent('ikLogout');
     };
 
     handleEmitUserGoToDashboard = (e: Event): void => {