+++ webapp/channels/src/components/post_view/post_list/index.tsx
@@ -11,6 +11,7 @@ import {getRecentPostsChunkInChannel, makeGetPostsChunkAroundPost, getUnreadPost
 import {memoizeResult} from 'mattermost-redux/utils/helpers';
 import {makePreparePostIdsForPostList} from 'mattermost-redux/utils/post_list';
 
+import {loadDeletedPosts} from 'actions/channel_actions';
 import {updateNewMessagesAtInChannel} from 'actions/global_actions';
 import {
     loadPosts,
@@ -112,6 +113,7 @@ function mapDispatchToProps(dispatch: Dispatch) {
             loadPosts,
             loadLatestPosts,
             loadPostsAround,
+            loadDeletedPosts,
             syncPostsInChannel,
             markChannelAsRead,
             updateNewMessagesAtInChannel,