+++ webapp/channels/src/packages/mattermost-redux/src/reducers/entities/teams.ts
@@ -11,6 +11,9 @@ import type {RelationOneToOne, IDMappedObjects} from '@mattermost/types/utilitie
 import {AdminTypes, ChannelTypes, TeamTypes, UserTypes, SchemeTypes, GroupTypes} from 'mattermost-redux/action_types';
 import {teamListToMap} from 'mattermost-redux/utils/team_utils';
 
+import {isServerVersionGreaterThanOrEqualTo} from 'utils/server_version';
+import {getDesktopVersion, isDesktopApp} from 'utils/user_agent';
+
 function currentTeamId(state = '', action: AnyAction) {
     switch (action.type) {
     case TeamTypes.SELECT_TEAM:
@@ -52,12 +55,26 @@ function teams(state: IDMappedObjects<Team> = {}, action: AnyAction) {
     case TeamTypes.UPDATED_TEAM:
     case TeamTypes.PATCHED_TEAM:
     case TeamTypes.REGENERATED_TEAM_INVITE_ID:
-    case TeamTypes.RECEIVED_TEAM:
-        return {
+    case TeamTypes.RECEIVED_TEAM: {
+        const newTeams = {
             ...state,
             [action.data.id]: action.data,
         };
 
+        if (isDesktopApp() && isServerVersionGreaterThanOrEqualTo(getDesktopVersion(), '3.2.0')) {
+            window.postMessage(
+                {
+                    type: 'update-teams',
+                    message: {
+                        teams: Object.values(newTeams),
+                    },
+                },
+                window.origin,
+            );
+        }
+
+        return newTeams;
+    }
     case TeamTypes.RECEIVED_TEAM_DELETED: {
         const nextState = {...state};
         const teamId = action.data.id;